test: move optimizer adapter into fuzz harness
